[Tab Switcher] Prevent merging collaborations via bulk edit
If multiple collaborations are selected in the multi-select UI disable
the ability to merge the groups.

When merging and a single collaboration is selected, the destination
group will be the collaboration as otherwise the collaboration will be
destroyed.

Bug: 362818090
Change-Id: I24bc38f7c9596b21b6aaf98068015a2c97613aa9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5854081
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Brandon Fong <bjfong@google.com>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354125}