Prevent multiple keys from signalling they are OSCrypt Sync compatible
There was a (test only) issue whereby it was possible to set
multiple keys that indicated they were OSCrypt Sync compatible.

This uncovered a bug in the .Clone() function because the keyring is
a std::map the assumption in the code that the keys were always
ordered was false. This caused Clone() to sometimes return a different
Encryptor when kEncryptSyncCompat parameter was supplied.

Multiple keys that are OSCrypt Sync compatible only happens in tests,
so a CHECK is added to enforce this, tests that did this are removed,
and a new death test is added to verify this can never happen.

BUG=40241934

Change-Id: I36c4a4610f2510751fc3829a9aed67c71d46926a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5855480
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354206}