[Extensions] Instantiate PermissionsManager factory early
KeyedService factories should be instantiated before profile
initialization is complete, and should normally be done in the
`EnsureCoreBrowserContextKeyedServiceFactoriesBuilt()` methods.
PermissionsManager wasn't -- in production, this didn't cause issues
because other classes called into PermissionsManager as part of their
setup, which resulted in implicitly creating the PermissionsManager.
However, this is an implementation detail, and not guaranteed to be
consistent (and isn't consistent in the experimental android-desktop
build config).

Register PermissionsManager's factory as part of the standard
`EnsureCoreBrowserContextKeyedServiceFactoriesBuilt()` method.

Bug: 356905053
Change-Id: I521120b078730a128f5224510c0b69baed95cced
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5855301
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Auto-Submit: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354232}