Implement Document-Policy: expect-no-embedded-resources
The expect-no-linked-resources configuration point in Document Policy allows a document to hint to the user agent to better optimize its loading sequence, such as not using the default speculative parsing behavior. The current implementation on this CL skips attaching PreloadScanner if the policy is supplied.

Bug: 365632977
Change-Id: Ia4b307a4eac2051cca77d6ad1fe8354aa74a6790
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5840722
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Alex N. Jose <alexnj@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354246}