Propagate CacheMode through redirect in MBDS
Previously in the MultiBufferDataSource fiefdom, UrlIndex::GetByUrl took
a CacheMode flag that was supposed to prevent requests from re-using a
multibuffer when devtools had set the page to no-cache mode. However, if
the request offered a redirect,
ResourceMultiBufferDataProvider::DidReceiveResponse was querying for a
new UrlData object with the kNormal cachemode, dropping the
kCacheDisabled flag.

R=dalecurtis

Change-Id: I8609eb83d9d6c3e7af1f24fce6e606cb827d7681
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5855806
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354281}