[ios] Create Create MagicStackCollectionViewCell
MVTs use MagicStackModuleContainer, which is a UICollectionViewCell
subclass, and it seems like the Discover UICollectionView has
awareness of it as a cell (see c#14 for logs that lead to this
conclusion. This is the reasoning behind the unexpected behavior seen in
the bug.

To fix this, MagicStackModuleCollectionViewCell is created as the
cell used in the Magic Stack, while MagicStackModuleContainer will be
used by MVTs and as the primary subview for
MagicStackModuleCollectionViewCell. Most of the code is just moving
necessary scoped logic (i.e. long-press context menu) to the cell
class and forwarding calls from the cell to the container. The
only changes are in MagicStackModuleCollectionViewCell's initWithFrame:,
where no longer sets self.contentView's corner radius and clipToBounds
since it does not seem to be necessary for UI consistency.

Video: https://drive.google.com/file/d/1kGMhbhYosTIyPvac5D12Cggkss7f7FD1/view?usp=sharing

Fixed: 360829555
Change-Id: Ic5d6b6ff3966329c809b7a355fc33925b3c767cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5855025
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1354298}