[Settings] Support standalone fragments
This patch defines two kinds of settings fragments: embeddable and
standalone. One can declare that a fragment is embeddable by
implementing the SettingsPage interface. Embeddable fragments are
"normal" fragments that can be shown within the multi-column UI.
Standalone fragments are always shown as a whole with a dedicated
activity, even if the single-activity mode is enabled.

This patch fixes the UI issues in the privacy guide page that is a
standalone fragment as it does not implement SettingsPage. Since the
page modifies the action bar to remove the up button, it used to not
work properly as embeddable fragments.

Bug: b/356743945
Change-Id: I796448b56697a64cd055356bd410b38ffe625d43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5848768
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sirisha Kavuluru <skavuluru@google.com>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354325}