Remove dangling WebContents on shutdown
Some crashes were found in the DWC of
WebContentsOfBrowserContext::~WebContentsOfBrowserContext.
It showed that in some cases, when the browser context is shut down
, the WebContent in extensions::WebAuthFlow is left hanging.

This CL adds an observer for the browser context in WebAuthFlow to help us reset the WebContent when the browser context is destroyed.

Bug: 358487779
Change-Id: I334010d9159616c094485c0bda020b6a4829ce75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5775488
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: q lamry <zhaoy@microsoft.com>
Reviewed-by: Leon Han <shulianghan@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1354339}