SelectFileDialogExtension: Fix termination to always cleanup
In rare occasions the dialog and its delegate have the termination
initiated by the aura::Window, instead of content::WebContents.
See both stack traces in the bug comment #13.

When this happens the views::WidgetDelegate::WindowWillClose() isn't
called. The SelectFileDialogExtension and its delegate
SystemFilesAppDialogDelegate relied on `WindowWillClose()` to perform
their cleanup. Thus, the cleanup wasn't performed causing a crash.

This CL changes the cleanup to be based on:
WebDialogDelegate::RegisterOnDialogClosedCallback() which fixes the
issue, this callback is triggered for both termination paths.

Bug: 361042814
Change-Id: Icde5174fb119860bccefebefaab059382ac10af8
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5832772
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354353}