arc: Introduce MojoInitData & Send the version of ArcBridgeHost
This CL adds a protocol to send the UUID and the version of each
interface during initialization. When the ARC side became capable of
handling this protocol, we can skip QueryVersion from ArcBridgeInstance
to Chrome and reduce an unnecessary IPC.

This CL introduces a class MojoInitData which contains data to be sent
during the bootstrap connection. When kArcExchangeVersionOnMojoHandshake
feature is enabled, it switches to the new protocol and additionally
provides the UUID and the version of each interface. Though in this CL
it only adds the information of ArcBridgeHost, the number of interfaces
will be increased later.

Note that this CL does not change the behavior of existing protocol
because all of the new features are under the
kArcExchangeVersionOnMojoHandshake flag.

Bug: b:363922359, b:355365971
Change-Id: Id41884705b1248e0f80fe21cbbc3155382917337
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5782514
Commit-Queue: Soshun Naito <soshun@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Eriko Kurimoto <elkurin@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354365}