Fix Data URL Length histogram
This CL fixes a typo in the histogram name to use uppercase instead of
lowercase, so it shows up on the UMA dashboard correctly.
This CL also extends the expiry of this histogram to allow for more
time to monitor the data.

OBSOLETE_HISTOGRAMS=Patterned histogram Blink.Network.DataURLLength{Type} is replaced by Blink.Network.DataUrlLength{Type}

Bug: 348442535
Change-Id: Iaa886a969d676c27497996c4f6448057a339a224
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5849644
Commit-Queue: Nidhi Jaju <nidhijaju@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354396}