Change Update to UpdateOrSet
The rationale for this change is that we now add recovery authfactor
on as soon as the user accepts T&C. So we are finding that users are
getting in a situation where they only have Recovery AuthFactor setup.
Ideally that is fine, and in a truly passwordless world that is the
expected scenario. Therefore when we go through something like that,
we either update or set.

BUG=b:328912970

Change-Id: I4f36528656c5c54cc62eb913e9dd40f0d6200eef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5757067
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354430}