[UNO] Refactor sign in promo password move to callback
Refactor the password move from local to account store after a sign in
through the password sign in promo to use callbacks instead of passing
the password form. This is done in preparation for the addition of
address sign in promos.

Bug: b:319411636

Change-Id: I1b8b444c179d1e72872514fda4695a6cee69dc47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5797300
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Amelie Schneider <amelies@google.com>
Reviewed-by: Ryan Sultanem <rsult@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1354454}