[Settings] Fix crash in the single-activity mode
SyncPromoPreference crashes when it is reattached to the preference
hierarchy, which happens in the single-activity mode where a single
instance of SettingsActivity is used to navigate settings. This is
because SyncPromoPreference#onDetach releases a reference to
SyncPromoController, while Preference in general can be attached to the
preference hierarchy multiple times.

Bug: b/356743945
Change-Id: If76bd1d12f42d69c04d10366d9bcf507c79bb383
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5856842
Reviewed-by: Samar Chehade-Lepleux <samarchehade@google.com>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354464}