Default to the TOPLEFT variants of the DXGI color space formats for BT.2020.
This seems to address the banding issues we were seeing in the
test case reported in the bug.

There are comments in this block which indicate that either of
LEFT or TOPLEFT DXGI formats for 2020 could be picked.

This seems to be related to the chroma location field which is displayed to be topleft in the test videos by ffprobe. chrome_location refers to the location of the chroma (color) samples relative to the luma (brightness) samples in a video frame.
The chroma samples (color information) are aligned with the top-left corner of the luma samples in the frame. In other words, the chroma samples are positioned directly at the same spatial coordinates as the corresponding luma samples. Using the LEFT DXGI formats could be causing the chroma color values to be interpreted differently leading to banding.

The presentation codepath in the DXGI swap chain does not get information on the chroma location in the color space format used by
the video. Hence using either LEFT or TOPLEFT is a heuristic at the moment.

Changes
1. Added a check for the FULL range in the BT.2020 color space check
   in the GetDXGIColorSpace() function. We return
   DXGI_COLOR_SPACE_YCBCR_FULL_G22_LEFT_P2020 in this case.

2. Default to TOPLEFT for the YUV limited cases for BT.2020.

3. In the SwapChainPresenter::VideoProcessorBlt() function call
   the context1->VideoProcessorSetOutputColorSpace1 function even
   if IDXGISwapChain3::SetColorSpace1 function fails. Left a comment
   in the code as to why this is. https://www.asawicki.info/news_1703_programming_hdr_monitor_support_in_direct3d

Impact
This change improves the visual quality of SDR 10-bit videos by
eliminating banding artifacts associated with incorrect color space
conversion

BUG=362787710

Change-Id: If8ad615d6dea72a38a6f89af4cfb5fb0bfc75e75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5824521
Reviewed-by: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Anantanarayanan Iyengar US <aiyengar@nvidia.com>
Cr-Commit-Position: refs/heads/main@{#1354476}