Implement try again for plus address error modals
For creation errors that are neither affiliation errors not quota
errors (e.g. time outs, identification issues, etc.), the user has
the option to "Try again" in the error modal that is shown. This
CL implements the logic for reshowing the popup in AED.

Follow-up CLs will customize the UI more and add tests for it.

Bug: 362445807, 363695012
Change-Id: If45cbe984d0cff5ff9247a79d3fad3a92ad8a46c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5857280
Commit-Queue: Jan Keitel <jkeitel@google.com>
Reviewed-by: Timofey Chudakov <tchudakov@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1354506}