[fuzzing] Don't terminate ipc_interfaces_dumper if it takes time
The ipc_interfaces_dumper executable, which is run via the
renderer_in_process_mojolpm_fuzzer_generator target in
chrome/test/fuzzing/renderer_fuzzing, is implemented as a browser
test. As such, running it will ordinarily use the test launcher to run
the test in a subprocess. In this particular case, we don't necessarily
want the launcher to terminate the test process if it happens to take a
while because in some configurations the dumper simply may have a lot of
work to do. So, pass --single-process-tests to the executable, thereby
instructing the test launcher to run the test directly in the main
process. This should resolve build failures that have a lot of output
that ends with something like:

[2/2] IPCInterfacesDumper.DumperTest (TIMED OUT)
1 test timed out:
    IPCInterfacesDumper.DumperTest (../../chrome/test/fuzzing/renderer_fuzzing/ipc_fuzzing/ipc_interfaces_dumper.cc:69)

Change-Id: I0b5ebbf268d90835edd6fcd3ab7508966779d4b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5857675
Reviewed-by: Titouan Rigoudy <titouan@chromium.org>
Auto-Submit: Greg Thompson <grt@chromium.org>
Commit-Queue: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354507}