[Android] Enhance documentation for payments UI info classes.
This CL is a follow up of the crrev.com/c/5826654 that addresses
comments to the original CL. The motivation of the original CL was to
enhance code readability by maintaining fewer states in the UI
information classes.

NO_IFTTT=Didn't add any new parameters to the constructor.

Fixed: 365730633
Change-Id: If6afb2e05378ebc70e2a160c24ad8ce0850f98f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5857163
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Timofey Chudakov <tchudakov@google.com>
Cr-Commit-Position: refs/heads/main@{#1354543}