[iOS][QD] Fix screen flash & add DUMP_WILL_BE_CHECK
This CL addresses two issues:

1. First, it adds a DUMP_WILL_BE_CHECK to allow us to better debug why
the bottom sheet momentarily comes back. Currently, we have the
hypothesis that perhaps the deletion can be triggered twice and that
might be causing the bug.

2. Second, it fixes a momentarily screen flash of the tab grid. The
screen flash in the tab grid was caused by the gradient for the grid
cell masks starting with toValue - byValue which in practise made them
transparent at the beginning.

Fixed: 365776279
Bug: 335387869, 354112735, 366161260
Change-Id: Iefc7faafe5886391257d7d663068c4ab69652685
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5851773
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Louis Romero <lpromero@google.com>
Commit-Queue: Filipa Senra <fsenra@google.com>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354551}