Enable sync checker to check allowlist on Clank
This is an optimization to improve the speed of Safe Browsing checks.
Only applying it on Clank because allowlist check is not significantly
faster than blocklist check on desktop platforms and the allowlist is
not available on WebView.

This feature is controlled by the SafeBrowsingSyncCheckerCheckAllowlist
flag. It is disabled by default.

See go/skip-sync-hpd-allowlist-android for details.

Bug: 347890373
Change-Id: I8c78f5a55677a154b4d8c424e58d60c46d149d7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5855364
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354572}