Implement `AndroidSensitiveContentClient` and set view sensitivity
The `AndroidSensitiveContentClient` is tied to one `WebContents`, and
owns the `SensitiveContentManager`.

The client serves as a bridge between the manager and the embedder.

In Clank, the client is instantiated by `TabHelpers::AttachTabHelpers`.
The manager decides when the content of the page becomes sensitive/not
sensitive, passes this information to the C++ implementation of the
client, which then passes this information to the Java implementation of
the client. The latter updates the content sensitivity of the container
view of its web contents.

WebView will use the same client, but that will be done in a separate
CL.

Design doc: go/sensitive-content-api-chrome

Low-Coverage-Reason: COVERAGE_UNDERREPORTED: The lack of coverage comes from the fact that `SensitiveContentClient.java` has 2 constructors, one of which is meant to be used only in production.
Bug: 343119998
Change-Id: I1b3b17b1716fe09fc3dd454f02b160034cd0622e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5823622
Reviewed-by: Victor Tan <victortan@chromium.org>
Commit-Queue: Theo Cristea <theocristea@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Reviewed-by: Friedrich Hauser <friedrichh@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354603}