[iOS] Add UIBlockerManagerObserver
When SigninPolicySceneAgent needs to show the sign-in dialog,
the dialog is skipped if there is a UI blocker.
Once the UI blocker is removed, SigninPolicySceneAgent needs to test
again if it is possible display the sign-in display.

This patch adds observers for `UIBlockerManager` and
`-[id<UIBlockerManagerObserver> currentUIBlockerRemoved]` as soon as
`-[id<UIBlockerManager> currentUIBlocker] is nil.

Bug: 365747792
Change-Id: Id4805460dfb5bb16b9b0bf419b4e9249928d7e05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5854863
Auto-Submit: Jérôme Lebel <jlebel@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354611}